Created CliBenchmarkRunner.cs — internal static class with a single RunAsync method taking the module, subcommand name, stripped options dict, CliOptions, and CancellationToken. Contains both the detailed (per-generation stats) and simple (total elapsed) paths. Program.cs — the 60-line benchmark block collapsed to three lines (benchOptions prep + await CliBenchmarkRunner.RunAsync(...) + return 0), and the now-unused using System.Diagnostics removed.
